Module google_bigquerydatatransfer1::api

Structs

BigQueryDataTransfer
Central instance to access all BigQueryDataTransfer related resource activities

CheckValidCredsRequest
A request to determine whether the user has valid credentials. This method is used to limit the number of OAuth popups in the user interface. The user id is inferred from the API call context. If the data source has the Google+ authorization type, this method returns false, as it cannot be determined whether the credentials are already valid merely based on the user id.

CheckValidCredsResponse
A response indicating whether the credentials exist and are valid.

DataSource
Represents data source metadata. Metadata is sufficient to render UI and request proper OAuth tokens.

DataSourceParameter
Represents a data source parameter with validation rules, so that parameters can be rendered in the UI. These parameters are given to us by supported data sources, and include all needed information for rendering and validation. Thus, whoever uses this api can decide to generate either generic ui, or custom data source specific forms.

EmailPreferences
Represents preferences for sending email notifications for transfer run events.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

ListDataSourcesResponse
Returns list of supported data sources and their metadata.

ListLocationsResponse
The response message for Locations.ListLocations.

ListTransferConfigsResponse
The returned list of pipelines in the project.

ListTransferLogsResponse
The returned list transfer run messages.

ListTransferRunsResponse
The returned list of pipelines in the project.

Location
A resource that represents Google Cloud Platform location.

ProjectDataSourceCheckValidCredCall
Returns true if valid credentials exist for the given data source and requesting user. Some data sources doesn't support service account, so we need to talk to them on behalf of the end user. This API just checks whether we have OAuth token for the particular user, which is a pre-requisite before user can create a transfer config.

ProjectDataSourceGetCall
Retrieves a supported data source and returns its settings, which can be used for UI rendering.

ProjectDataSourceListCall
Lists supported data sources and returns their settings, which can be used for UI rendering.

ProjectLocationDataSourceCheckValidCredCall
Returns true if valid credentials exist for the given data source and requesting user. Some data sources doesn't support service account, so we need to talk to them on behalf of the end user. This API just checks whether we have OAuth token for the particular user, which is a pre-requisite before user can create a transfer config.

ProjectLocationDataSourceGetCall
Retrieves a supported data source and returns its settings, which can be used for UI rendering.

ProjectLocationDataSourceListCall
Lists supported data sources and returns their settings, which can be used for UI rendering.

ProjectLocationGetCall
Gets information about a location.

ProjectLocationListCall
Lists information about the supported locations for this service.

ProjectLocationTransferConfigCreateCall
Creates a new data transfer configuration.

ProjectLocationTransferConfigDeleteCall
Deletes a data transfer configuration, including any associated transfer runs and logs.

ProjectLocationTransferConfigGetCall
Returns information about a data transfer config.

ProjectLocationTransferConfigListCall
Returns information about all data transfers in the project.

ProjectLocationTransferConfigPatchCall
Updates a data transfer configuration. All fields must be set, even if they are not updated.

ProjectLocationTransferConfigRunDeleteCall
Deletes the specified transfer run.

ProjectLocationTransferConfigRunGetCall
Returns information about the particular transfer run.

ProjectLocationTransferConfigRunListCall
Returns information about running and completed jobs.

ProjectLocationTransferConfigRunTransferLogListCall
Returns user facing log messages for the data transfer run.

ProjectLocationTransferConfigScheduleRunCall
Creates transfer runs for a time range [start_time, end_time]. For each date - or whatever granularity the data source supports - in the range, one transfer run is created. Note that runs are created per UTC time in the time range. DEPRECATED: use StartManualTransferRuns instead.

ProjectLocationTransferConfigStartManualRunCall
Start manual transfer runs to be executed now with schedule_time equal to current time. The transfer runs can be created for a time range where the run_time is between start_time (inclusive) and end_time (exclusive), or for a specific run_time.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the BigQueryDataTransfer hub.

ProjectTransferConfigCreateCall
Creates a new data transfer configuration.

ProjectTransferConfigDeleteCall
Deletes a data transfer configuration, including any associated transfer runs and logs.

ProjectTransferConfigGetCall
Returns information about a data transfer config.

ProjectTransferConfigListCall
Returns information about all data transfers in the project.

ProjectTransferConfigPatchCall
Updates a data transfer configuration. All fields must be set, even if they are not updated.

ProjectTransferConfigRunDeleteCall
Deletes the specified transfer run.

ProjectTransferConfigRunGetCall
Returns information about the particular transfer run.

ProjectTransferConfigRunListCall
Returns information about running and completed jobs.

ProjectTransferConfigRunTransferLogListCall
Returns user facing log messages for the data transfer run.

ProjectTransferConfigScheduleRunCall
Creates transfer runs for a time range [start_time, end_time]. For each date - or whatever granularity the data source supports - in the range, one transfer run is created. Note that runs are created per UTC time in the time range. DEPRECATED: use StartManualTransferRuns instead.

ProjectTransferConfigStartManualRunCall
Start manual transfer runs to be executed now with schedule_time equal to current time. The transfer runs can be created for a time range where the run_time is between start_time (inclusive) and end_time (exclusive), or for a specific run_time.

ScheduleOptions
Options customizing the data transfer schedule.

ScheduleTransferRunsRequest
A request to schedule transfer runs for a time range.

ScheduleTransferRunsResponse
A response to schedule transfer runs for a time range.

StartManualTransferRunsRequest
A request to start manual transfer runs.

StartManualTransferRunsResponse
A response to start manual transfer runs.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

TimeRange
A specification for a time range, this will request transfer runs with run_time between start_time (inclusive) and end_time (exclusive).

TransferConfig
Represents a data transfer configuration. A transfer configuration contains all metadata needed to perform a data transfer. For example, destination_dataset_id specifies where data should be stored. When a new transfer configuration is created, the specified destination_dataset_id is created when needed and shared with the appropriate data source service account.

TransferMessage
Represents a user facing message for a particular data transfer run.

TransferRun
Represents a data transfer run.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.