Struct google_bigquerydatatransfer1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the BigQueryDataTransfer hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_bigquerydatatransfer1 as bigquerydatatransfer1;
 
use std::default::Default;
use oauth2;
use bigquerydatatransfer1::BigQueryDataTransfer;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = BigQueryDataTransfer::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `data_sources_check_valid_creds(...)`, `data_sources_get(...)`, `data_sources_list(...)`, `locations_data_sources_check_valid_creds(...)`, `locations_data_sources_get(...)`, `locations_data_sources_list(...)`, `locations_get(...)`, `locations_list(...)`, `locations_transfer_configs_create(...)`, `locations_transfer_configs_delete(...)`, `locations_transfer_configs_get(...)`, `locations_transfer_configs_list(...)`, `locations_transfer_configs_patch(...)`, `locations_transfer_configs_runs_delete(...)`, `locations_transfer_configs_runs_get(...)`, `locations_transfer_configs_runs_list(...)`, `locations_transfer_configs_runs_transfer_logs_list(...)`, `locations_transfer_configs_schedule_runs(...)`, `locations_transfer_configs_start_manual_runs(...)`, `transfer_configs_create(...)`, `transfer_configs_delete(...)`, `transfer_configs_get(...)`, `transfer_configs_list(...)`, `transfer_configs_patch(...)`, `transfer_configs_runs_delete(...)`, `transfer_configs_runs_get(...)`, `transfer_configs_runs_list(...)`, `transfer_configs_runs_transfer_logs_list(...)`, `transfer_configs_schedule_runs(...)` and `transfer_configs_start_manual_runs(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn data_sources_check_valid_creds(
&self,
request: CheckValidCredsRequest,
name: &str
) -> ProjectDataSourceCheckValidCredCall<'a, C>

Create a builder to help you perform the following task:

Returns true if valid credentials exist for the given data source and requesting user. Some data sources doesn't support service account, so we need to talk to them on behalf of the end user. This API just checks whether we have OAuth token for the particular user, which is a pre-requisite before user can create a transfer config.

Arguments

request - No description provided.

name - Required. The data source in the form: projects/{project_id}/dataSources/{data_source_id} or projects/{project_id}/locations/{location_id}/dataSources/{data_source_id}.

pub fn data_sources_get(&self, name: &str) -> ProjectDataSourceGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a supported data source and returns its settings, which can be used for UI rendering.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/dataSources/{data_source_id} or projects/{project_id}/locations/{location_id}/dataSources/{data_source_id}

pub fn data_sources_list(
&self,
parent: &str
) -> ProjectDataSourceListCall<'a, C>

Create a builder to help you perform the following task:

Lists supported data sources and returns their settings, which can be used for UI rendering.

Arguments

parent - Required. The BigQuery project id for which data sources should be returned. Must be in the form: projects/{project_id} or `projects/{project_id}/locations/{location_id}

pub fn locations_data_sources_check_valid_creds(
&self,
request: CheckValidCredsRequest,
name: &str
) -> ProjectLocationDataSourceCheckValidCredCall<'a, C>

Create a builder to help you perform the following task:

Returns true if valid credentials exist for the given data source and requesting user. Some data sources doesn't support service account, so we need to talk to them on behalf of the end user. This API just checks whether we have OAuth token for the particular user, which is a pre-requisite before user can create a transfer config.

Arguments

request - No description provided.

name - Required. The data source in the form: projects/{project_id}/dataSources/{data_source_id} or projects/{project_id}/locations/{location_id}/dataSources/{data_source_id}.

pub fn locations_data_sources_get(
&self,
name: &str
) -> ProjectLocationDataSourceGetCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a supported data source and returns its settings, which can be used for UI rendering.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/dataSources/{data_source_id} or projects/{project_id}/locations/{location_id}/dataSources/{data_source_id}

pub fn locations_data_sources_list(
&self,
parent: &str
) -> ProjectLocationDataSourceListCall<'a, C>

Create a builder to help you perform the following task:

Lists supported data sources and returns their settings, which can be used for UI rendering.

Arguments

parent - Required. The BigQuery project id for which data sources should be returned. Must be in the form: projects/{project_id} or `projects/{project_id}/locations/{location_id}

pub fn locations_transfer_configs_runs_transfer_logs_list(
&self,
parent: &str
) -> ProjectLocationTransferConfigRunTransferLogListCall<'a, C>

Create a builder to help you perform the following task:

Returns user facing log messages for the data transfer run.

Arguments

parent - Required. Transfer run name in the form: projects/{project_id}/transferConfigs/{config_id}/runs/{run_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}/runs/{run_id}

pub fn locations_transfer_configs_runs_delete(
&self,
name: &str
) -> ProjectLocationTransferConfigRunDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified transfer run.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id}/runs/{run_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}/runs/{run_id}

pub fn locations_transfer_configs_runs_get(
&self,
name: &str
) -> ProjectLocationTransferConfigRunGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about the particular transfer run.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id}/runs/{run_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}/runs/{run_id}

pub fn locations_transfer_configs_runs_list(
&self,
parent: &str
) -> ProjectLocationTransferConfigRunListCall<'a, C>

Create a builder to help you perform the following task:

Returns information about running and completed jobs.

Arguments

parent - Required. Name of transfer configuration for which transfer runs should be retrieved. Format of transfer configuration resource name is: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}.

pub fn locations_transfer_configs_create(
&self,
request: TransferConfig,
parent: &str
) -> ProjectLocationTransferConfigCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new data transfer configuration.

Arguments

request - No description provided.

parent - Required. The BigQuery project id where the transfer configuration should be created. Must be in the format projects/{project_id}/locations/{location_id} or projects/{project_id}. If specified location and location of the destination bigquery dataset do not match - the request will fail.

pub fn locations_transfer_configs_delete(
&self,
name: &str
) -> ProjectLocationTransferConfigDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a data transfer configuration, including any associated transfer runs and logs.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}

pub fn locations_transfer_configs_get(
&self,
name: &str
) -> ProjectLocationTransferConfigGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about a data transfer config.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}

pub fn locations_transfer_configs_list(
&self,
parent: &str
) -> ProjectLocationTransferConfigListCall<'a, C>

Create a builder to help you perform the following task:

Returns information about all data transfers in the project.

Arguments

parent - Required. The BigQuery project id for which data sources should be returned: projects/{project_id} or projects/{project_id}/locations/{location_id}

pub fn locations_transfer_configs_patch(
&self,
request: TransferConfig,
name: &str
) -> ProjectLocationTransferConfigPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a data transfer configuration. All fields must be set, even if they are not updated.

Arguments

request - No description provided.

name - The resource name of the transfer config. Transfer config names have the form projects/{project_id}/locations/{region}/transferConfigs/{config_id}. Where config_id is usually a uuid, even though it is not guaranteed or required. The name is ignored when creating a transfer config.

pub fn locations_transfer_configs_schedule_runs(
&self,
request: ScheduleTransferRunsRequest,
parent: &str
) -> ProjectLocationTransferConfigScheduleRunCall<'a, C>

Create a builder to help you perform the following task:

Creates transfer runs for a time range [start_time, end_time]. For each date - or whatever granularity the data source supports - in the range, one transfer run is created. Note that runs are created per UTC time in the time range. DEPRECATED: use StartManualTransferRuns instead.

Arguments

request - No description provided.

parent - Required. Transfer configuration name in the form: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}.

pub fn locations_transfer_configs_start_manual_runs(
&self,
request: StartManualTransferRunsRequest,
parent: &str
) -> ProjectLocationTransferConfigStartManualRunCall<'a, C>

Create a builder to help you perform the following task:

Start manual transfer runs to be executed now with schedule_time equal to current time. The transfer runs can be created for a time range where the run_time is between start_time (inclusive) and end_time (exclusive), or for a specific run_time.

Arguments

request - No description provided.

parent - Transfer configuration name in the form: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.

pub fn transfer_configs_runs_transfer_logs_list(
&self,
parent: &str
) -> ProjectTransferConfigRunTransferLogListCall<'a, C>

Create a builder to help you perform the following task:

Returns user facing log messages for the data transfer run.

Arguments

parent - Required. Transfer run name in the form: projects/{project_id}/transferConfigs/{config_id}/runs/{run_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}/runs/{run_id}

pub fn transfer_configs_runs_delete(
&self,
name: &str
) -> ProjectTransferConfigRunDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified transfer run.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id}/runs/{run_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}/runs/{run_id}

pub fn transfer_configs_runs_get(
&self,
name: &str
) -> ProjectTransferConfigRunGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about the particular transfer run.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id}/runs/{run_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}/runs/{run_id}

pub fn transfer_configs_runs_list(
&self,
parent: &str
) -> ProjectTransferConfigRunListCall<'a, C>

Create a builder to help you perform the following task:

Returns information about running and completed jobs.

Arguments

parent - Required. Name of transfer configuration for which transfer runs should be retrieved. Format of transfer configuration resource name is: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}.

pub fn transfer_configs_create(
&self,
request: TransferConfig,
parent: &str
) -> ProjectTransferConfigCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new data transfer configuration.

Arguments

request - No description provided.

parent - Required. The BigQuery project id where the transfer configuration should be created. Must be in the format projects/{project_id}/locations/{location_id} or projects/{project_id}. If specified location and location of the destination bigquery dataset do not match - the request will fail.

pub fn transfer_configs_delete(
&self,
name: &str
) -> ProjectTransferConfigDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a data transfer configuration, including any associated transfer runs and logs.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}

pub fn transfer_configs_get(
&self,
name: &str
) -> ProjectTransferConfigGetCall<'a, C>

Create a builder to help you perform the following task:

Returns information about a data transfer config.

Arguments

name - Required. The field will contain name of the resource requested, for example: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}

pub fn transfer_configs_list(
&self,
parent: &str
) -> ProjectTransferConfigListCall<'a, C>

Create a builder to help you perform the following task:

Returns information about all data transfers in the project.

Arguments

parent - Required. The BigQuery project id for which data sources should be returned: projects/{project_id} or projects/{project_id}/locations/{location_id}

pub fn transfer_configs_patch(
&self,
request: TransferConfig,
name: &str
) -> ProjectTransferConfigPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a data transfer configuration. All fields must be set, even if they are not updated.

Arguments

request - No description provided.

name - The resource name of the transfer config. Transfer config names have the form projects/{project_id}/locations/{region}/transferConfigs/{config_id}. Where config_id is usually a uuid, even though it is not guaranteed or required. The name is ignored when creating a transfer config.

pub fn transfer_configs_schedule_runs(
&self,
request: ScheduleTransferRunsRequest,
parent: &str
) -> ProjectTransferConfigScheduleRunCall<'a, C>

Create a builder to help you perform the following task:

Creates transfer runs for a time range [start_time, end_time]. For each date - or whatever granularity the data source supports - in the range, one transfer run is created. Note that runs are created per UTC time in the time range. DEPRECATED: use StartManualTransferRuns instead.

Arguments

request - No description provided.

parent - Required. Transfer configuration name in the form: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}.

pub fn transfer_configs_start_manual_runs(
&self,
request: StartManualTransferRunsRequest,
parent: &str
) -> ProjectTransferConfigStartManualRunCall<'a, C>

Create a builder to help you perform the following task:

Start manual transfer runs to be executed now with schedule_time equal to current time. The transfer runs can be created for a time range where the run_time is between start_time (inclusive) and end_time (exclusive), or for a specific run_time.

Arguments

request - No description provided.

parent - Transfer configuration name in the form: projects/{project_id}/transferConfigs/{config_id} or projects/{project_id}/locations/{location_id}/transferConfigs/{config_id}.